backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
drf-spectacular==0.27.1
celery==5.3.6
redis==5.0.1
django-filter==23.5
django-prometheus==2.3.1
sqlalchemy==2.0.27
django-allauth==0.61.1
dj-rest-auth==5.0.2
django-redis==5.4.0 
sqlalchemy
google-generativeai==0.8.3